DOUBLE PLS

+++ zuul-framework/src/TWoT_A1/TWoT.java
@@ -919,7 +919,7 @@ public class TWoT implements Serializable {
      *
      * @return
      */
-    public int getPlayerHealth() {
+    public double getPlayerHealth() {
         return player.getHealth();
     }
